Entity and DTO updated

+++ src/main/java/lk/ijse/meethive/dto/EventDTO.java
@@ -16,7 +16,7 @@ import java.util.List;
 @Getter
 @Setter
 public class EventDTO {
-    private Long eventId;
+    private int eventId;
     private String eventTitle;
     private String description;
     private LocalDateTime startDateTime;
