+++ src/main/java/lk/ijse/meethive/entity/Role.java
@@ -16,7 +16,7 @@ import java.util.List;
 public class Role {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long roleId;
+    private int roleId;
 
     private String role;
